ai_backend/requirements.txt
fastapi
pydantic
transformers
faiss-cpu
mysql-connector-python
uvicorn